Group tasks by model
